esio: added esio_file_clone method #1167
Implementation, documentation, and unit tests

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/esio/trunk@16349 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ tests/basic.c
@@ -33,6 +33,7 @@
 #include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include <unistd.h>
 #include <mpi.h>
 #include <hdf5.h>
@@ -50,14 +51,33 @@
 #pragma warning(pop)
 #endif
 
+// Add command line options
+static const fctcl_init_t my_cl_options[] = {
+    {
+        "--preserve",
+        "-t",
+        FCTCL_STORE_TRUE,
+        "Are temporary filenames displayed and files preserved?"
+    },
+    FCTCL_INIT_NULL /* Sentinel */
+};
+
 FCT_BGN()
 {
+    int preserve = 0;
+
     // MPI setup: MPI_Init and atexit(MPI_Finalize)
     int world_size, world_rank;
     MPI_Init(&argc, &argv);
     atexit((void (*) ()) MPI_Finalize);
     MPI_Comm_size(MPI_COMM_WORLD, &world_size);
     MPI_Comm_rank(MPI_COMM_WORLD, &world_rank);
+
+    // Install the command line options defined above.
+    fctcl_install(my_cl_options);
+
+    // Retrieve options
+    preserve = fctcl_is("--preserve");
 
     // Obtain default HDF5 error handler
     H5E_auto2_t hdf5_handler;
@@ -79,11 +99,31 @@ FCT_BGN()
     {
         FCT_SETUP_BGN()
         {
-            (void) input_dir; // Unused
+            ESIO_MPICHKQ(MPI_Barrier(MPI_COMM_WORLD)); // Synchronize
+
+            // Restore HDF5/ESIO default error handling
             H5Eset_auto2(H5E_DEFAULT, hdf5_handler, hdf5_client_data);
             esio_set_error_handler(esio_handler);
-            filename = create_testfilename(filetemplate);
+
+            // Rank 0 generates a unique filename and broadcasts it
+            int filenamelen;
+            if (world_rank == 0) {
+                filename = create_testfilename(filetemplate);
+                if (preserve) {
+                    printf("\nfilename: %s\n", filename);
+                }
+                filenamelen = strlen(filename);
+            }
+            ESIO_MPICHKR(MPI_Bcast(&filenamelen, 1, MPI_INT,
+                                   0, MPI_COMM_WORLD));
+            if (world_rank > 0) {
+                filename = calloc(filenamelen + 1, sizeof(char));
+            }
+            ESIO_MPICHKR(MPI_Bcast(filename, filenamelen, MPI_CHAR,
+                                    0, MPI_COMM_WORLD));
             assert(filename);
+
+            // Initialize ESIO state
             state = esio_handle_initialize(MPI_COMM_WORLD);
             assert(state);
         }
@@ -91,8 +131,14 @@ FCT_BGN()
 
         FCT_TEARDOWN_BGN()
         {
-            if (filename) free(filename);
+            // Finalize ESIO state
             esio_handle_finalize(state);
+
+            // Clean up the unique file and filename
+            if (world_rank == 0) {
+                if (!preserve) unlink(filename);
+            }
+            if (filename) free(filename);
         }
         FCT_TEARDOWN_END();
 
@@ -127,7 +173,11 @@ FCT_BGN()
             H5Eset_auto2(H5E_DEFAULT, hdf5_handler, hdf5_client_data);
 
             // Remove the file and create without overwrite should succeed
-            fct_req(0 == unlink(filename));
+            ESIO_MPICHKQ(MPI_Barrier(MPI_COMM_WORLD)); // Synchronize
+            if (world_rank == 0) {
+                fct_req(0 == unlink(filename));
+            }
+            ESIO_MPICHKQ(MPI_Barrier(MPI_COMM_WORLD)); // Synchronize
             fct_req(0 == esio_file_create(state, filename, 0 /* no overwrite */));
 
             // Close the file
@@ -144,9 +194,48 @@ FCT_BGN()
 
             // Close the file
             fct_req(0 == esio_file_close(state));
+        }
+        FCT_TEST_END();
 
-            // Remove the file
-            fct_req(0 == unlink(filename));
+        FCT_TEST_BGN(file_clone)
+        {
+            // Dynamically create the empty.h5 source filename per input_dir
+            fct_req(NULL != input_dir /* check ESIO_TEST_INPUT_DIR set */);
+            int srcfilelen = strlen(input_dir);
+            fct_req(srcfilelen > 0);
+            srcfilelen += strlen("/empty.h5");
+            srcfilelen += 1;
+            char * srcfile = calloc(srcfilelen, 1);
+            fct_req(NULL != srcfile);
+            fct_req(NULL != strcat(srcfile, input_dir));
+            fct_req(NULL != strcat(srcfile, "/empty.h5"));
+
+            // Clone with overwrite should always work
+            fct_req(0 == esio_file_clone(state, srcfile, filename, 1));
+
+            // Close the file
+            fct_req(0 == esio_file_close(state));
+
+            // Create without overwrite should fail
+            H5Eset_auto(H5E_DEFAULT, NULL, NULL);
+            esio_set_error_handler_off();
+            fct_req(0 != esio_file_clone(state, srcfile, filename, 0));
+            esio_set_error_handler(esio_handler);
+            H5Eset_auto2(H5E_DEFAULT, hdf5_handler, hdf5_client_data);
+
+            // Remove the file and create without overwrite should succeed
+            ESIO_MPICHKQ(MPI_Barrier(MPI_COMM_WORLD)); // Synchronize
+            if (world_rank == 0) {
+                fct_req(0 == unlink(filename));
+            }
+            ESIO_MPICHKQ(MPI_Barrier(MPI_COMM_WORLD)); // Synchronize
+            fct_req(0 == esio_file_clone(state, srcfile, filename, 0));
+
+            // Close the file
+            fct_req(0 == esio_file_close(state));
+
+            // Deallocate the srcfile name
+            if (srcfile) free(srcfile);
         }
         FCT_TEST_END();
